Move top sort test into test_sort
